+++ packages/react/src/contexts/authContext.tsx
@@ -1,10 +1,12 @@
 import { ApolloError } from '@apollo/client'
+import { ConnectExtension } from '@magic-ext/connect'
 import { MeemAPI, makeFetcher } from '@meemproject/sdk'
 import WalletConnectProvider from '@walletconnect/web3-provider'
 import { providers, ethers } from 'ethers'
 import Cookies from 'js-cookie'
 import JWT from 'jsonwebtoken'
 import { DateTime } from 'luxon'
+import { Magic } from 'magic-sdk'
 import React, {
 	createContext,
 	useMemo,
@@ -107,6 +109,12 @@ export interface IAuthContextState {
 
 	/** The current user */
 	me?: MeemAPI.v1.GetMe.IResponseBody
+
+	/** The type of wallet that is connected */
+	walletType?: WalletType
+
+	/** An instance of the Magic.link sdk instance */
+	magic?: Magic<ConnectExtension[]>
 }
 
 const AuthContext = createContext({} as IAuthContextState)
@@ -114,6 +122,10 @@ AuthContext.displayName = 'AuthContext'
 
 export interface IAuthContextProps {
 	children?: ReactNode
+
+	chainId: number
+
+	magicApiKey: string
 
 	/** Use custom RPC endpoints for a chain */
 	rpcs?: {
@@ -128,16 +140,31 @@ export interface IAuthContextProps {
 	}
 }
 
+// export type WalletType = 'magic' | 'injected' | 'wallet' | undefined
+
+export type WalletType =
+	| 'magic'
+	| 'metamask'
+	| 'coinbase_wallet'
+	| 'wallet_connect'
+	| 'injected'
+	| 'walletconnect'
+	| undefined
+
 export const AuthProvider: React.FC<IAuthContextProps> = ({
 	rpcs,
+	chainId: initialChainId,
+	magicApiKey,
 	...props
 }: IAuthContextProps) => {
 	const [accounts, setAccounts] = useState<string[]>([])
 	const [jwt, setJwt] = useState<string>()
-	const [chainId, setChainId] = useState<number>()
+	const [magic, setMagic] = useState<Magic<ConnectExtension[]>>()
+	const [web3Modal, setWeb3Modal] = useState<Web3Modal | undefined>(undefined)
+	const [walletType, setWalletType] = useState<WalletType>()
+	const [chainId, setChainId] = useState<number>(initialChainId)
 	const [loginState, setLoginState] = useState<LoginState>(LoginState.Unknown)
 	const [isConnected, setIsConnected] = useState<boolean>(false)
-	const [web3Modal, setWeb3Modal] = useState<Web3Modal | undefined>(undefined)
 	const [provider, setProvider] = useState<any | undefined>(undefined)
 	const [web3Provider, setWeb3Provider] = useState<
 		providers.Web3Provider | undefined
@@ -199,6 +226,12 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 
 	useEffect(() => {
 		const meemJwtToken = Cookies.get('meemJwtToken')
+		const wt = Cookies.get('walletType')
+
+		if (wt) {
+			log.debug('setting wallet type from cookie', { wt })
+			setWalletType(wt as WalletType)
+		}
 
 		if (meemJwtToken) {
 			const result = JWT.decode(meemJwtToken) as Record<string, any>
@@ -218,24 +251,54 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 	const connectWallet = useCallback(
 		async (to?: string) => {
 			let p: any
+			let w3p: any
 
-			try {
-				// This is the initial `provider` that is returned when
-				// using web3Modal to connect. Can be MetaMask or WalletConnect.
-				p = to ? await web3Modal?.connectTo(to) : await web3Modal?.connect()
-				log.debug('Connected')
-			} catch (e) {
-				log.warn(e)
-				log.warn(
-					'Error: unable to connect with web3modal. Did the user close the modal?'
-				)
+			if (to === 'walletconnect' || to === 'injected') {
+				try {
+					// This is the initial `provider` that is returned when
+					// using web3Modal to connect. Can be MetaMask or WalletConnect.
+					p = to ? await web3Modal?.connectTo(to) : await web3Modal?.connect()
+					setWalletType(to)
+					w3p = new providers.Web3Provider(p, 'any')
+					log.debug('Connected', { provider: p })
+				} catch (e) {
+					log.warn(e)
+					log.warn(
+						'Error: unable to connect with web3modal. Did the user close the modal?'
+					)
+					return
+				}
+			} else if (walletType === 'magic' || to === 'magic') {
+				try {
+					const customNodeOptions = {
+						rpcUrl:
+							chainId && rpcUrls[chainId]
+								? (rpcUrls[chainId][0] as string)
+								: '',
+						chainId
+					}
+
+					const m = new Magic(magicApiKey, {
+						network: customNodeOptions,
+						extensions: [new ConnectExtension()]
+					})
+
+					// @ts-ignore
+					w3p = new ethers.providers.Web3Provider(m.rpcProvider)
+
+					setMagic(m)
+					log.debug('Connected via magic')
+				} catch (e) {
+					log.warn(e)
+					log.warn('Error: unable to connect with magic link.')
+					return
+				}
+			} else {
+				log.warn('Unable to determine wallet type to connect')
 				return
 			}
 
-			// We plug the initial `provider` into ethers.js and get back
-			// a Web3Provider. This will add on methods from ethers.js and
-			// event listeners such as `.on()` will be different.
-			const w3p = new providers.Web3Provider(p, 'any')
+			// const p = new ethers.providers.Web3Provider(m.rpcProvider)
 
 			const s = w3p.getSigner()
 
@@ -252,8 +315,25 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 			setWeb3Provider(w3p)
 			setIsConnected(true)
 		},
-		[web3Modal]
+		[chainId, web3Modal, walletType]
 	)
+
+	useEffect(() => {
+		const go = async () => {
+			const walletInfo = await magic?.connect.getWalletInfo()
+			if (walletInfo) {
+				Cookies.set('walletType', walletInfo.walletType, {
+					secure:
+						typeof window !== 'undefined' &&
+						window.location.protocol === 'https:'
+				})
+
+				setWalletType(walletInfo?.walletType)
+			}
+		}
+
+		go()
+	}, [magic])
 
 	useEffect(() => {
 		if (typeof window !== 'undefined') {
@@ -276,9 +356,15 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 	// Auto connect to the cached provider
 	useEffect(() => {
 		if (web3Modal?.cachedProvider) {
-			connectWallet()
+			connectWallet('injected')
 		}
 	}, [connectWallet, web3Modal])
+
+	useEffect(() => {
+		if (meData && walletType) {
+			connectWallet()
+		}
+	}, [connectWallet, meData])
 
 	const handleAccountsChanged = useCallback((acc: string[]) => {
 		setAccounts(acc)
@@ -305,15 +391,20 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 
 	const disconnectWallet = useCallback(async () => {
 		setMeemJwt(undefined)
-		web3Modal?.clearCachedProvider()
+
 		if (provider?.disconnect && typeof provider.disconnect === 'function') {
 			await provider.disconnect()
 		}
 
+		Cookies.remove('walletType')
+
+		web3Modal?.clearCachedProvider()
+
+		magic?.connect.disconnect()
 		setWeb3Provider(undefined)
 		setAccounts([])
 		meMutate()
-	}, [provider, web3Modal])
+	}, [provider, magic, web3Modal])
 
 	const handleChainChanged = useCallback(
 		(chainHex: string) => {
@@ -327,6 +418,7 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 		if (network?.chainId === newChainId) {
 			return
 		}
+		// @ts-ignore
 		const { ethereum } = window
 
 		const chain = chains.find(c => +c.chainId === +newChainId)
@@ -363,6 +455,7 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 			}
 
 			try {
+				// @ts-ignore
 				await window.ethereum.request({
 					method: 'wallet_switchEthereumChain',
 					params: [{ chainId: formattedChainId }]
@@ -421,7 +514,9 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 			jwt,
 			setChain,
 			chainId,
-			me: meData
+			me: meData,
+			walletType,
+			magic
 		}),
 		[
 			web3Provider,
@@ -438,7 +533,9 @@ export const AuthProvider: React.FC<IAuthContextProps> = ({
 			jwt,
 			setChain,
 			chainId,
-			meData
+			meData,
+			walletType,
+			magic
 		]
 	)