+++ packages/sdk/src/gql/agreement.gql.ts
@@ -1,4 +1,4 @@
-import { gql } from '@apollo/client'
+import { gql } from '@apollo/client/core'
 
 export const GET_CONTRACT_QUERY = gql`
 	query GetContract($contractId: uuid!) {